+++ widgets/lib/buttons/buttons.dart
@@ -1,0 +1,11 @@
+/// A library for buttons
+/// That include Text button, Parallalogram Button and Icon Button with all of thier variations
+///
+
+library buttons;
+
+export 'button_style.dart';
+export 'icon_button.dart';
+export 'parallalogram_button.dart';
+export 'text_button.dart';
+export 'button_enums.dart';