+++ widgets/lib/switches/switches.dart
@@ -1,0 +1,5 @@
+/// A library to import and use switches/toggles
+library switches;
+
+export 'woi_switch_button.dart';
+export 'switch_style_model.dart';