+++ widgets/lib/woi_widgets.dart
@@ -1,8 +1,12 @@
-library buttons;
+/// The all in one library to access all of the widgets
 
-/// A button library files
-export 'buttons/icon_button.dart';
-export 'buttons/text_button.dart';
-export 'buttons/button_style.dart';
-export 'buttons/parallalogram_button.dart';
-export 'utils/enums.dart';
+library widgets;
+
+/// A library for Buttons including Text, Parallelogram and Icon button
+export 'buttons/buttons.dart';
+
+/// Library for Switches
+export 'switches/switches.dart';
+
+/// Library for radio buttons and checkmarks
+export 'radio_and_checkbox/radio_and_checkbox.dart';