+++ widgets/CHANGELOG.md
@@ -9,3 +9,12 @@
 
 ## ^0.0.3
 * Update and fixed documentation while adding code on how to use the widgets.
+
+## ^0.0.4
+* New widgets are here! -> Switches/toggles, Radio Button and Checkbox.
+
+## ^0.0.5
+* Minor fix in the navigation of radio button and checkbox documentation in readme.
+
+## ^0.0.6
+* Updated and improved documentation.